build shared library using cmake under windows and linux, added pom.xml to compile the samples, fixed the notes from the pull request
