+++ c/build_all/linux/build.sh
@@ -13,7 +13,7 @@ build_amqp=ON
 build_http=ON
 build_mqtt=ON
 skip_unittests=OFF
-run_valgrind=
+build_javawrapper=OFF
 
 usage ()
 {
@@ -29,7 +29,7 @@ usage ()
     echo " --no-http                     do no build HTTP transport and samples"
     echo " --no-mqtt                     do no build MQTT transport and samples"
     echo " --toolchain-file <file>       pass cmake a toolchain file for cross compiling"
-    echo " -rv, --run_valgrind           will execute ctest with valgrind"
+    echo " --build-javawrapper           build java c wrapper module"
     exit 1
 }
 
@@ -60,8 +60,8 @@ process_args ()
               "--no-amqp" ) build_amqp=OFF;;
               "--no-http" ) build_http=OFF;;
               "--no-mqtt" ) build_mqtt=OFF;;
+	      "--build-javawrapper" ) build_javawrapper=ON;;
               "--toolchain-file" ) save_next_arg=2;;
-              "-rv" | "--run_valgrind" ) run_valgrind=1;;
               * ) usage;;
           esac
       fi
@@ -79,14 +79,7 @@ process_args $*
 rm -r -f ~/cmake
 mkdir ~/cmake
 pushd ~/cmake
-cmake $toolchainfile -DcompileOption_C:STRING="$extracloptions" -Drun_e2e_tests:BOOL=$run_e2e_tests -Drun_longhaul_tests=$run_longhaul_tests -Duse_amqp:BOOL=$build_amqp -Duse_http:BOOL=$build_http -Duse_mqtt:BOOL=$build_mqtt -Dskip_unittests:BOOL=$skip_unittests $build_root
+cmake $toolchainfile -DcompileOption_C:STRING="$extracloptions" -Drun_e2e_tests:BOOL=$run_e2e_tests -Drun_longhaul_tests=$run_longhaul_tests -Duse_amqp:BOOL=$build_amqp -Duse_http:BOOL=$build_http -Duse_mqtt:BOOL=$build_mqtt -Dskip_unittests:BOOL=$skip_unittests -Dbuild_javawrapper:BOOL=$build_javawrapper $build_root
 make --jobs=$(nproc)
 ctest -C "Debug" -V
-
-if [[ $run_valgrind == 1 ]] ;
-then
-	ctest -j $(nproc) -D ExperimentalMemCheck -VV
-
-fi
-
 popd
