+++ javawrapper/build_all/linux/exportLibPath.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# Copyright (c) Microsoft. All rights reserved.
+# Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+echo set the env variable LD_LIBRARY_PATH
+export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:~/cmake/javawrapper/
+echo $LD_LIBRARY_PATH
