+++ javawrapper/device/iothub_client_javawrapper/windows/dllmain.c
@@ -1,0 +1,22 @@
+// Copyright(c) Microsoft.All rights reserved.
+// Licensed under the MIT license.See LICENSE file in the project root for full license information.
+
+// dllmain.c : Defines the entry point for the DLL application.
+#include <windows.h>
+
+BOOL APIENTRY DllMain( HMODULE hModule,
+                       DWORD  ul_reason_for_call,
+                       LPVOID lpReserved
+					 )
+{
+	switch (ul_reason_for_call)
+	{
+	case DLL_PROCESS_ATTACH:
+	case DLL_THREAD_ATTACH:
+	case DLL_THREAD_DETACH:
+	case DLL_PROCESS_DETACH:
+		break;
+	}
+	return TRUE;
+}
+
