--- javawrapper/device/samples/send-event/readme.md
@@ -1,7 +1,0 @@
-
-# Run sample for windows
-
-- open the Visual Studio project iot_hub_client_dll in \javawrapper\windows and compile it.
-- create an eclipse project and add Send.java and all the .java files in \iothub-client-javawrapper
-- add the jna.jar library
-- copy in the java directory project all the .dll files obtained from the iot_hub_client_dll project
