--- javawrapper/windows/packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Apache.QPID.Proton.AzureIot" version="0.9.0.1-preview-003" targetFramework="native" />
-  <package id="Microsoft.Azure.C.SharedUtility" version="1.0.0-preview-003" targetFramework="native" />
-  <package id="Microsoft.Azure.IoTHub.AmqpTransport" version="1.0.0-preview-011" targetFramework="native" />
-  <package id="Microsoft.Azure.IoTHub.HttpTransport" version="1.0.0-preview-011" targetFramework="native" />
-  <package id="Microsoft.Azure.IoTHub.IoTHubClient" version="1.0.0-preview-011" targetFramework="native" />
-  <package id="Microsoft.Azure.IoTHub.MqttTransport" version="1.0.0-preview-011" targetFramework="native" />
-  <package id="Microsoft.Azure.IoTHub.Serializer" version="1.0.0-preview-011" targetFramework="native" />
-</packages>
